US20600: Introducing flash operations support
A set of flash and mode control operations added to the chr
device and avaialble via IOCTL. That includes flash read/write/erase,
checksum check and setting/getting of the current mode.

Some of the operations reports that the work is finished via
interrupts - standard one or by setting and signalling the error.
As the core device handles the errors as another type of interrupt then
the chr driver uses handles both interrupts to get the operation result.

Also minor cleanup/fixes for the previous change are included.