Added reboot hook module
The reboot hook module responsiblity is to register kernel hooks for
reboot and poweroff and notify the EFI PMM that the device should
be resarted or powered off.

NOTE: The reboot hook module requires atomic operations in the i2c
driver to be supported (see e.g. recent changes to imx-i2c) as well
as the proper hooks symbols being exposed by the kernel.